Move JetpackFullPluginInstallOnboardingUiStateMapper to the correct package
